Add readmes for each subfolder
This might seem too much, but will prove useful for new contributors if we add to these readme files whenever we think of something useful. For example if there is something to know specifically about the layouts, the readme file that is in their subfolder should be used for that purpose.